Move edit profile into component
