+++ app/Components/EditVolunteerForm/EditVolunteerFormInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace SousedskaPomoc\Components;
+
+use SousedskaPomoc\Components\EditVolunteerFormControl;
+
+interface IEditVolunteerFormInterface
+{
+    /** @return EditVolunteerFormControl */
+    public function create();
+}
