+++ app/Components/EditVolunteerForm/edit.latte
@@ -1,0 +1,1 @@
+{control editVolunteerForm}
